Extract vega renderer into avenger-vega package
