+++ examples/avenger-wasm-vega/bootstrap.js
