+++ examples/avenger-wasm-vega/webpack.config.js
