+++ javascript/avenger-vega/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "avenger-vega",
+  "version": "0.1.0",
+  "description": "",
+  "main": "src/index.js",
+  "files": [
+    "src/**/*.js"
+  ],
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "dependencies": {
+    "avenger-wasm": "../../avenger-wasm/pkg"
+  },
+  "peerDependencies": {
+    "vega-scenegraph": "^4.11.2",
+    "vega-util": "^1.17.2"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
